integration_testパッケージを使用したインテグレーションテストに移行
- Flutter標準のintegration_testパッケージを使用
- test/integrationからintegration_test/にリファクタ
- ダイビング記録数を全レコード数でカウントするように変更
- 底部ナビゲーションのテストを統合・拡張
- テストドライバーとREADMEを追加
