--- test/integration/bottom_navigation_test.dart
@@ -1,201 +1,0 @@
-import 'package:dive_log_book/features/statistics/statistics_screen.dart';
-import 'package:dive_log_book/main.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/semantics.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:sqflite_common_ffi/sqflite_ffi.dart';
-
-void main() {
-  late Widget app;
-
-  setUpAll(() {
-    // sqlfiteの初期化
-    sqfliteFfiInit();
-    databaseFactory = databaseFactoryFfi;
-  });
-
-  setUp(() async {
-    app = const ProviderScope(child: MyApp());
-  });
-
-  group('Bottom Navigation Integration Test', () {
-    testWidgets('底部ナビゲーションに統計タブが表示される', (WidgetTester tester) async {
-      // This test MUST fail until bottom navigation is implemented
-
-      expect(() async {
-        await tester.pumpWidget(app);
-        await tester.pumpAndSettle();
-
-        // 底部ナビゲーションバーの存在確認
-        expect(find.byType(BottomNavigationBar), findsOneWidget);
-
-        // 統計タブの存在確認
-        expect(find.text('統計'), findsOneWidget);
-        expect(find.byIcon(Icons.analytics), findsOneWidget);
-      }, throwsA(isA<Error>())); // 底部ナビゲーションが未実装
-    });
-
-    testWidgets('統計タブをタップして統計画面に遷移できる', (WidgetTester tester) async {
-      // This test MUST fail until navigation is implemented
-
-      expect(() async {
-        await tester.pumpWidget(app);
-        await tester.pumpAndSettle();
-
-        // 統計タブをタップ
-        await tester.tap(find.text('統計'));
-        await tester.pumpAndSettle();
-
-        // 統計画面が表示されることを確認
-        expect(find.byType(StatisticsScreen), findsOneWidget);
-        expect(find.text('総潜水時間'), findsOneWidget);
-      }, throwsA(isA<Error>())); // ナビゲーションが未実装
-    });
-
-    testWidgets('他のタブから統計タブへの遷移が正常に動作する', (WidgetTester tester) async {
-      // This test MUST fail until navigation is implemented
-
-      expect(() async {
-        await tester.pumpWidget(app);
-        await tester.pumpAndSettle();
-
-        // 初期状態（ダイビングログリスト画面と仮定）の確認
-        // expect(find.byType(DiveLogListScreen), findsOneWidget);
-
-        // 統計タブに遷移
-        await tester.tap(find.text('統計'));
-        await tester.pumpAndSettle();
-
-        // 統計画面が表示されることを確認
-        expect(find.byType(StatisticsScreen), findsOneWidget);
-
-        // 元の画面に戻る
-        await tester.tap(find.text('ログ')); // 仮のタブ名
-        await tester.pumpAndSettle();
-
-        // 元の画面に戻っていることを確認
-        // expect(find.byType(DiveLogListScreen), findsOneWidget);
-      }, throwsA(isA<Error>())); // ナビゲーションが未実装
-    });
-
-    testWidgets('底部ナビゲーションの選択状態が正しく更新される', (WidgetTester tester) async {
-      // This test MUST fail until navigation state management is implemented
-
-      expect(() async {
-        await tester.pumpWidget(app);
-        await tester.pumpAndSettle();
-
-        // 底部ナビゲーションバーを取得
-        final bottomNav = tester.widget<BottomNavigationBar>(
-          find.byType(BottomNavigationBar),
-        );
-
-        // 初期選択状態の確認（通常は0番目）
-        expect(bottomNav.currentIndex, 0);
-
-        // 統計タブをタップ
-        await tester.tap(find.text('統計'));
-        await tester.pumpAndSettle();
-
-        // 選択状態が更新されていることを確認
-        final updatedBottomNav = tester.widget<BottomNavigationBar>(
-          find.byType(BottomNavigationBar),
-        );
-        expect(updatedBottomNav.currentIndex, isNot(0)); // 統計タブのインデックス
-      }, throwsA(isA<Error>())); // ナビゲーション状態管理が未実装
-    });
-
-    testWidgets('Material Design 3のナビゲーションスタイルが適用されている', (
-      WidgetTester tester,
-    ) async {
-      // This test MUST fail until navigation styling is implemented
-
-      expect(() async {
-        await tester.pumpWidget(app);
-        await tester.pumpAndSettle();
-
-        // Material Design 3のナビゲーションバーの確認
-        final bottomNav = tester.widget<BottomNavigationBar>(
-          find.byType(BottomNavigationBar),
-        );
-
-        // Material 3のスタイル属性確認
-        expect(bottomNav.type, BottomNavigationBarType.fixed);
-        // expect(bottomNav.useLegacyColorScheme, false); // Material 3対応
-      }, throwsA(isA<Error>())); // ナビゲーションスタイリングが未実装
-    });
-
-    testWidgets('統計画面のAppBarタイトルが正しく表示される', (WidgetTester tester) async {
-      // This test MUST fail until AppBar integration is implemented
-
-      expect(() async {
-        await tester.pumpWidget(app);
-        await tester.pumpAndSettle();
-
-        // 統計タブをタップ
-        await tester.tap(find.text('統計'));
-        await tester.pumpAndSettle();
-
-        // AppBarのタイトル確認
-        expect(find.text('統計'), findsWidgets);
-        // または
-        expect(find.text('ダイビング統計'), findsOneWidget);
-      }, throwsA(isA<Error>())); // AppBar統合が未実装
-    });
-
-    testWidgets('デバイスの戻るボタンが正しく処理される', (WidgetTester tester) async {
-      // This test MUST fail until navigation handling is implemented
-
-      expect(() async {
-        await tester.pumpWidget(app);
-        await tester.pumpAndSettle();
-
-        // 統計画面に遷移
-        await tester.tap(find.text('統計'));
-        await tester.pumpAndSettle();
-
-        // デバイスの戻るボタンをシミュレート
-        await tester.binding.defaultBinaryMessenger.handlePlatformMessage(
-          'flutter/navigation',
-          null,
-          (data) {},
-        );
-        await tester.pumpAndSettle();
-
-        // アプリが適切に処理されることを確認
-        // (具体的な動作は設計次第 - 前の画面に戻るか、アプリ終了か)
-      }, throwsA(isA<Error>())); // ナビゲーション処理が未実装
-    });
-
-    testWidgets('アクセシビリティ対応が適切に設定されている', (WidgetTester tester) async {
-      // This test MUST fail until accessibility is implemented
-
-      expect(() async {
-        await tester.pumpWidget(app);
-        await tester.pumpAndSettle();
-
-        // セマンティクス情報の確認
-        expect(find.bySemanticsLabel('統計画面を開く'), findsOneWidget);
-        expect(find.bySemanticsLabel('ダイビング記録画面を開く'), findsOneWidget);
-
-        // タブの選択状態がセマンティクスで正しく報告される
-        final statisticsTab = tester.getSemantics(find.text('統計'));
-        expect(
-          statisticsTab.hasFlag(SemanticsFlag.isSelected),
-          isFalse,
-        ); // 初期状態
-
-        // 統計タブをタップ
-        await tester.tap(find.text('統計'));
-        await tester.pumpAndSettle();
-
-        final selectedStatisticsTab = tester.getSemantics(find.text('統計'));
-        expect(
-          selectedStatisticsTab.hasFlag(SemanticsFlag.isSelected),
-          isTrue,
-        ); // 選択状態
-      }, throwsA(isA<Error>())); // アクセシビリティが未実装
-    });
-  });
-}
